Support different env files for dev and prod
In the past the Postgres password was set using the .env file in the
project main folder. Variables set there could be reused
in the docker compose file in the "environment" section. Veriables set
in .env files that are specified using the "env_file" in the compose
file itself can't be directly accessed in the "environment" section
of the compose file, see https://stackoverflow.com/a/51527492/166229
That's why I now fetch the POSTGRES_PASSWORD in the Django file
directly.